feat: add holidays functionality implemented
