+++ apps/api/src/app/time-off-request/time-off-request.entity.ts
@@ -4,7 +4,8 @@ import {
 	JoinColumn,
 	ManyToMany,
 	JoinTable,
-	OneToOne
+	OneToOne,
+	ManyToOne
 } from 'typeorm';
 import { Base } from '../core/entities/base';
 import { TimeOff as ITimeOffRequest, StatusTypesEnum } from '@gauzy/models';
@@ -15,13 +16,8 @@ import { TimeOffPolicy } from '../time-off-policy/time-off-policy.entity';
 
 @Entity('time_off_request')
 export class TimeOffRequest extends Base implements ITimeOffRequest {
-	@ApiProperty({ type: String })
-	@IsString()
-	@IsOptional()
-	@Column()
-	holidayName?: string;
 
-	@ManyToMany((type) => Employee, { cascade: ['update'] })
+	@ManyToMany((type) => Employee, { cascade: true })
 	@JoinTable({
 		name: 'time_off_request_employee'
 	})
@@ -40,7 +36,7 @@ export class TimeOffRequest extends Base implements ITimeOffRequest {
 
 	@ApiProperty({ type: TimeOffPolicy })
 	@IsOptional()
-	@OneToOne((type) => TimeOffPolicy, { nullable: false, onDelete: 'CASCADE' })
+	@ManyToOne((type) => TimeOffPolicy, { nullable: false, onDelete: 'CASCADE' })
 	@JoinColumn()
 	policy?: TimeOffPolicy;
