+++ apps/api/src/app/time-off-request/time-off-request.module.ts
@@ -4,11 +4,14 @@ import { TimeOffRequestService } from './time-off-request.service';
 import { TimeOffRequest } from './time-off-request.entity';
 import { Employee } from '../employee/employee.entity';
 import { TimeOffRequestControler } from './time-off-request.controller';
+import { TimeOffPolicy } from '../time-off-policy/time-off-policy.entity';
+import { User } from '../user/user.entity';
+import { UserService } from '../user/user.service';
 
 @Module({
-	imports: [TypeOrmModule.forFeature([TimeOffRequest, Employee])],
+	imports: [TypeOrmModule.forFeature([TimeOffRequest, Employee, User, TimeOffPolicy])],
 	controllers: [TimeOffRequestControler],
-	providers: [TimeOffRequestService],
+	providers: [TimeOffRequestService, UserService],
 	exports: [TypeOrmModule]
 })
 export class TimeOffRequestModule {}
