+++ apps/gauzy/src/app/@shared/time-off/time-off-mutation.module.ts
@@ -18,7 +18,7 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { ThemeModule } from '../../@theme/theme.module';
 import { TimeOffSettingsMutationComponent } from './settings-mutation/time-off-settings-mutation.component';
 import { EmployeeSelectorsModule } from '../../@theme/components/header/selectors/employee/employee.module';
-import { TimeOffRequestMutationComponent } from './time-off-request-mutation/time-off-request--mutation.component';
+import { TimeOffRequestMutationComponent } from './time-off-request-mutation/time-off-request-mutation.component';
 
 export function HttpLoaderFactory(http: HttpClient) {
 	return new TranslateHttpLoader(http, './assets/i18n/', '.json');
